iris-dataset-pytorch/requirements.txt
scikit-learn
pytorch